+++ common/src/main/java/com/iamkaf/konfig/impl/v1/client/control/BaseSliderWidget.java
@@ -9,6 +9,7 @@ import static com.iamkaf.konfig.impl.v1.client.screen.KonfigScreenSupport.text;
 
 import com.iamkaf.konfig.impl.v1.client.screen.KonfigScreenMetrics;
 import net.minecraft.client.gui.components.AbstractSliderButton;
+import net.minecraft.client.input.MouseButtonEvent;
 import net.minecraft.util.Mth;
 
 @ApiStatus.Internal
@@ -20,6 +21,18 @@ public abstract class BaseSliderWidget extends AbstractSliderButton {
     public final void syncToProgress(double progress) {
         this.value = Mth.clamp(progress, 0.0D, 1.0D);
         this.updateMessage();
+    }
+
+    @Override
+    public boolean mouseClicked(MouseButtonEvent event, boolean doubleClick) {
+        if (event.button() == 1 && this.isActive() && this.isMouseOver(event.x(), event.y())) {
+            return this.resetToDefault();
+        }
+        return super.mouseClicked(event, doubleClick);
+    }
+
+    protected boolean resetToDefault() {
+        return false;
     }
 }
 //?}